tools/pyproject.toml
[project]
name = "ancient-kingdoms-tools"
version = "0.1.0"
description = "Python tools for Ancient Kingdoms data mining"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[tool.uv]
dev-dependencies = []
